Show Sales only if user is logged in
Sales are now retreived from Shopstyle, but not Amazon because
Amazon parent products do not have discounts associated with them. Only
the child products do. Thus, Amazon sales will be revisited in the future.

If a user is not logged in, sales will be hidden. If they are logged in
and a current user exists, sale prices are shown next to the existing retail price.