Implement Product Filtering by Category and Brand for Shopstyle
The networking architecture and Search Implentation have been heavily
modified to support filtering. The actual categories and brands that are displayed to the user are very limited and need to be updated. However, from this small set, the filtering works correctly, including integration with existing pagination logic and invalidating old filters.